yolo_api/requirements.txt
uvicorn
fastapi
pydantic
pydantic-settings
python-dotenv
ultralytics
pillow
python-multipart
boto3